move stuff into new disk.c
