+++ setup.py
@@ -208,6 +208,7 @@ elif FREEBSD:
             'psutil/_psutil_bsd.c',
             'psutil/arch/freebsd/cpu.c',
             'psutil/arch/freebsd/mem.c',
+            'psutil/arch/freebsd/disk.c',
             'psutil/arch/freebsd/specific.c',
             'psutil/arch/freebsd/sys_socks.c',
             'psutil/arch/freebsd/proc_socks.c',
